requirements.txt
cloudscraper
demjson3
docling
dotenv
duckduckgo_search
flask
flask-cors
googlesearch-python
ipython
langchain
langchain_core
langgraph
langchain-openai
langchain-pymupdf4llm
lxml
lxml_html_clean
markdownify
numpy
pandas
playwright
PyMuPDF
pymupdf4llm
PyPDF2
pytest
readability-lxml
requests
scikit-learn
selenium
sentence_transformers
tavily-python
urllib3
uvicorn
webdriver-manager
